+++ tests/templates/default/slotbox.php
@@ -1,0 +1,1 @@
+<div class="box"><?php $this->slot(); ?></div>