+++ tests/templates/default/slotdefault.php
@@ -1,0 +1,1 @@
+<div><?php if ($this->hasSlot()): ?><?php $this->slot(); ?><?php else: ?>fallback<?php endif; ?></div>