+++ tests/templates/default/slotonce.php
@@ -1,0 +1,3 @@
+<?php $this->insert('slotbox', slot: function (): void { ?>
+<b>hello</b>
+<?php }); ?>